51-frontend feature: Moved routing to different indepentend components
