+++ mobile/routers/Router.tsx
@@ -1,0 +1,6 @@
+import { View } from 'react-native';
+import MainRouter from './components/MainRouter';
+const Router = () => {
+  return <MainRouter />;
+};
+export default Router;
